Test: merge two test files to one.
